fixed some bus

+++ Web/Tasks/CRUD/view.js
@@ -3,19 +3,9 @@ class Page {
         this.target = target;
         this.noteArea = document.createElement('div');
         this.titlesArea = document.createElement('div');
-        this.note = new Note('Untitled', '');
+        this.target.append(this.noteArea, this.titlesArea);
+        this.note = null;
         this.notebook = new Notebook();
-        this.notebook.addNote(this.note);
-
-        let element = document.createElement('textarea');
-        element.setAttribute('id', 'noteTitle');
-        this.noteArea.append(element);
-        element = document.createElement('textarea');
-        element.setAttribute('id', 'noteBody');
-        this.noteArea.append(element);
-
-        this.target.append(this.noteArea);
-        this.target.append(this.titlesArea);
 
         this.newNote = this.newNote.bind(this);
         this.updateNoteBody = this.updateNoteBody.bind(this);
@@ -24,19 +14,41 @@ class Page {
         this.updateNoteTitle = this.updateNoteTitle.bind(this);
         this.drawNewButton = this.drawNewButton.bind(this);
         this.drawDeleteButton = this.drawDeleteButton.bind(this);
+        this.clearNoteArea = this.clearNoteArea.bind(this);
+        this.updateNoteBody = this.updateNoteBody.bind(this);
+        this.updateNoteTitle = this.updateNoteTitle.bind(this);
+        this.loadNote = this.loadNote.bind(this);
+        this.deleteAllNotes = this.deleteAllNotes.bind(this);
+        this.deleteNote = this.deleteNote.bind(this);
+
+        this.notebook.load().then(() => {
+            this.drawNotebook();
+        });
     }
 
     updateNoteTitle() {
         this.note.updateTitle(document.getElementById('noteTitle').value);
+        this.notebook.updateNote(this.note);
+        this.drawNotebook();
     }
 
     updateNoteBody() {
         this.note.updateBody(document.getElementById('noteBody').value);
+        this.notebook.updateNote(this.note);
     }
 
-    static clearNoteArea() {
-        document.getElementById('noteTitle').getEventListeners().change.forEach(event => { event.remove(); });
-        document.getElementById('noteBody').getEventListeners().change.forEach(event => { event.remove(); });
+    clearNoteArea() {
+        const title = document.getElementById('noteTitle');
+        const body = document.getElementById('noteBody');
+
+        if (title) {
+            title.removeEventListener('change', this.updateNoteTitle);
+            title.remove();
+        }
+        if (body) {
+            body.removeEventListener('click', this.updateNoteBody);
+            body.remove();
+        }
     }
 
     drawButton(id, callbackFunc) {
@@ -54,7 +66,9 @@ class Page {
     }
 
     drawNewButton() {
-        this.drawButton('new-btn', this.newNote).innerHTML = 'New Note';
+        if(!document.getElementById('new-btn')) {
+            this.drawButton('new-btn', this.newNote).innerHTML = 'New Note';
+        }
     }
 
     removeNewButton() {
@@ -62,7 +76,9 @@ class Page {
     }
 
     drawDeleteButton() {
-        this.drawButton('del-btn', this.deleteNote).innerHTML = 'Delete Note';
+        if(!document.getElementById('del-btn')) {
+            this.drawButton('del-btn', this.deleteNote).innerHTML = 'Delete Note';
+        }
     }
 
     removeDeleteButton() {
@@ -70,14 +86,16 @@ class Page {
     }
 
     drawDeleteAllButton() {
-        this.drawButton('delAll-btn', this.deleteAllNotes).innerHTML = 'Delete All Notes';
+        if(!document.getElementById('delAll-btn')) {
+            this.drawButton('delAll-btn', this.deleteAllNotes).innerHTML = 'Delete All Notes';
+        }  
     }
 
     removeDeleteAllButton() {
         Page.removeButton('delAll-btn', this.deleteAllNotes);
     }
 
-    clearTitlesArea() {
+    clearListArea() {
         this.titlesArea.childNodes.forEach(node => {
             node.removeEventListener('click', this.loadNote);
         });
@@ -85,7 +103,7 @@ class Page {
     }
 
     drawNote(noteID) {
-        Page.clearNoteArea();
+        this.clearNoteArea();
         this.note = this.notebook.notes[noteID];
 
         const titleElement = document.createElement('textarea');
@@ -94,33 +112,59 @@ class Page {
         titleElement.setAttribute('id', 'noteTitle');
         bodyElement.setAttribute('id', 'noteBody');
 
+        titleElement.addEventListener('input', this.updateNoteTitle);
+        bodyElement.addEventListener('input', this.updateNoteBody);
+
         titleElement.value = this.note.title;
-        bodyElement.valuse = this.note.body;
+        bodyElement.value = this.note.body;
+
+        this.noteArea.append(titleElement, bodyElement);
+        this.drawDeleteButton();
     }
 
     drawNotebook() {
-        this.clearTitlesArea();
-
-        notesAPI.getNotes().then(notes => {
-            notes.forEach(note => {
-                const title = document.createElement('li');
-                title.dataset.id = note.id;
-                title.addEventListener('click', this.loadNote);
-                title.innerHTML = note.title;
-                this.titlesArea.append(title);
-            });
+        this.clearListArea();
+        Object.keys(this.notebook.notes).forEach(key => {
+            const title = document.createElement('li');
+            title.dataset.id = key;
+            title.addEventListener('click', this.loadNote);
+            title.innerHTML = this.notebook.notes[key].title;
+            this.titlesArea.append(title);
         });
+        if (Object.keys(this.notebook.notes).length != 0 ) {
+            this.drawDeleteAllButton();
+        }
     }
 
-    loadNote() {
-        Page.clearNoteArea();
-        this.drawNote(this.dataset.id);
+    loadNote(element) {
+        this.clearNoteArea();
+        this.drawNote(element.currentTarget.dataset.id);
+        this.drawDeleteButton();
     }
 
-
-    newNote () {
+    newNote() {
         const note = new Note('Untitled', ' ');
-        note.update();
+        this.notebook.addNote(note);
+        this.drawNotebook();
         this.drawNote(note.ID);
+    }
+
+    deleteNote() {
+        this.clearNoteArea();
+        this.notebook.deleteNote(this.note.ID);
+        delete this.note;
+        this.removeDeleteButton();
+        this.removeDeleteAllButton();
+        this.drawNotebook();
+        
+    }
+
+    deleteAllNotes() {
+        this.clearListArea();
+        this.clearNoteArea();
+        this.notebook.deleteAll();
+        delete this.note;
+        this.removeDeleteAllButton();
+        this.removeDeleteButton();
     }
 }
